test to make sure assert fires in correct cycle
